requirements.txt
# ─── DRCode – AI Resume Analyzer ───
# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Telegram Bot
python-telegram-bot>=21.0

# File Parsing
PyMuPDF>=1.24.0
python-docx>=1.1.0

# NLP
spacy>=3.7.0

# Semantic Similarity
sentence-transformers>=2.5.0

# Resume Generation
reportlab>=4.1.0

# AI Improver (optional)
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.6.0
aiofiles>=23.2.1
httpx>=0.27.0

# MongoDB
motor>=3.3.0

# Authentication
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
